# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loyalty points & marketplace platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get the currently signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # --- Products ---
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List recent transactions for the current user
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [purchase, transfer, exchange] }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /transactions/purchase:
    post:
      operationId: createPurchase
      tags: [transactions]
      summary: Buy an item or coupon
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseBody" }
      responses:
        "201":
          description: Purchase confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400":
          description: Invalid purchase
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /transactions/transfer:
    post:
      operationId: createTransfer
      tags: [transactions]
      summary: Transfer loyalty points to another user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferBody" }
      responses:
        "201":
          description: Transfer complete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400":
          description: Invalid transfer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # --- Notifications ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role: { type: string, enum: [user, admin] }
        pointsBalance: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, role, pointsBalance, createdAt]
    UpdateUserBody:
      type: object
      properties:
        name: { type: string }
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
...
    Transaction:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [purchase, transfer, exchange] }
        buyerId:
          type: ["integer", "null"]
        buyerName:
          type: ["string", "null"]
        sellerId:
          type: ["integer", "null"]
        sellerName:
          type: ["string", "null"]
        productId:
          type: ["integer", "null"]
        productTitle:
          type: ["string", "null"]
        productImageUrl:
          type: ["string", "null"]
        pointsAmount: { type: integer }
        cashCents:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
        status: { type: string, enum: [pending, completed, failed] }
        createdAt: { type: string, format: date-time }
      required: [id, type, pointsAmount, status, createdAt]
    PurchaseBody:
      type: object
      properties:
        productId: { type: integer }
      required: [productId]
    TransferBody:
      type: object
      properties:
        recipientId: { type: integer }
        pointsAmount: { type: integer }
        note:
          type: ["string", "null"]
      required: [recipientId, pointsAmount]
...